Started work to increase ammount of useful TSDF data in published message
ROS imposes a limit of 1 GB on all messages. The TSDF could be up to 2 GB (on my current computer) so I'd like to export the full volume in a single message. The TSDF contains many empty voxels, which don't need to be exported in detail. A more clever message could replace many empty array entries with a single entry describing how many subsequent entries are also empty.